backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
numpy==1.26.1
pydantic==2.4.2
python-multipart==0.0.6
pillow==10.1.0
tensorflow==2.14.0
scikit-learn==1.3.2
matplotlib==3.8.1
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23 